Milestone 1 Phase 2/deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
plotly
Pillow
numpy
tensorflow
scikit-learn==1.2.2